Crate game_of_life

Structs

Grid

Struct that holds info with information and state of each cell

Indexes

Struct that holds info with Entity ID’s of each cell

Materials

Color Keeper

Position

Position struct

Run

Struct that holds information of pausing the game

SingleTile

Struct used to hold information of an individual cell

Size

Size struct xd

Tile

Sprite Struct used to spawn cells

Enums

State

Schrodinger Enum

Constants

ARENA_HEIGHT

ARENA_WIDTH

SCREEN_HEIGHT

SCREEN_WIDTH

Functions

change_state

Function that changes state with a mouse click We’re basically playing Gods now.

convert

Converts absolute coordinates into game coordinates (row, col)

get_index

Gets ArrayIndex based on Position Struct

initialize_grid

Initialize grid with all the dead cells

living_neighbours

Returns 2D Array with info of all the living neighbour cells

main

Main function and app builder Here the windows is defined and system are called

next_turn

Function that is called every one second that calculates all the future states and implements them (If game is not paused)

position_translation

Translats positions into 3D Vector

read_coords

Converts real mouse input into game coordinates

respawn

Kills grid and respawns it

setup

Setup function Here all the resources are initialized

size_scaling

Changes size of sprites

space_hit

Pauses and Repauses Game run is a parameter witch indicates the current state

spawn_grid

Originally spawning the grid