Struct game_of_life::state::Life

pub struct Life {
    x_size: usize,
    y_size: usize,
    grid: Vec<Vec<bool>>,
    blocks: (char, char),
}

Strcuture containing game state

Fields
x_size: usizey_size: usizegrid: Vec<Vec<bool>>blocks: (char, char)
Implementations

impl Life

Game implementation

pub fn new(x: u16, y: u16) -> Self

Creates new grid and fills it with dead or alive cells at random

fn populate(x_size: usize, y_size: usize) -> Vec<Vec<bool>>

Creates a 2D vector of random booleans

fn coords_to_flip(&self) -> Vec<(usize, usize)>

Returns a vector of tuples (y,x) which contains all coordinates on grid to be flipped

fn count_neighbors(&self, y: usize, x: usize) -> u8

Returns the number of cells that neighbor the cell at the given coordinates

pub fn turn(&mut self)

Executes one ‘turn’, flipping coordinates according to the Game of Life rules

pub fn draw(&self)

Clears terminal then draws grid